Backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
starlette>=0.27.0
typing-extensions>=4.8.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pymysql>=1.1.0
cryptography>=41.0.0